requirements.txt
discord.py>=2.6.4
python-dotenv>=1.1.1
yt-dlp>=2025.10.14
psutil>=7.1.0
pynacl>=1.6.0
aiohttp>=3.13.0
requests>=2.32.5